avoid warning
git-svn-id: https://svn.dealii.org/trunk@28713 0785d39b-7218-0410-832d-ea1e28bc413d

+++ deal.II/include/deal.II/lac/vector.templates.h
@@ -645,7 +645,7 @@ namespace internal
                 }
               n_chunks++;
             }
-          AssertDimension(X - X_original, vec_size);
+          AssertDimension(static_cast<unsigned int>(X - X_original), vec_size);
 
           // now sum the results from the chunks
           // recursively
